Move measurements.dart into core/theme
